modif SC , pages messages contact

+++ PrestaChope/tools/SuperController.php
@@ -33,11 +33,14 @@ Class SuperController {
 
 
             case "deconnexion":
-                require_once('pages/déconnexion/ControllerDéconnexion.php');
-                $deconnexion = new ControllerDéconnexion();
+                if (!empty($_SESSION['ID'])) {
+                    require_once('pages/déconnexion/ControllerDéconnexion.php');
+                    $deconnexion = new ControllerDéconnexion();
 
-                $deconnexion->déconnexion();
-                Rooter::redirectToPage('accueil');
+                    $deconnexion->déconnexion();
+                }
+                Rooter::redirectToPage("accueil");
+
                 break;
 
 
@@ -58,14 +61,18 @@ Class SuperController {
 
 
             case "contact":
-                require_once('pages/contact/ControllerContact.php');
-                $contacter = new ControllerContact();
+                if (!empty($_SESSION['ID'])) {
+                    require_once('pages/contact/ControllerContact.php');
+                    $contacter = new ControllerContact();
 
-                $contacter->includeViewContact();
+                    $contacter->includeViewContact();
 
-                if (!empty($_POST['contenu']) && !empty($_POST['type']) && !empty($_POST['idclient'])) {
-                    $contacter->insertContactDemande($_POST['contenu'], $_POST['type'], $_POST['idclient']);
-                    Rooter::redirectToPage('accueil');
+                    if (!empty($_POST['contenu']) && !empty($_POST['type']) && !empty($_POST['idclient'])) {
+                        $contacter->insertContactDemande($_POST['contenu'], $_POST['type'], $_POST['idclient']);
+                        Rooter::redirectToPage('accueil');
+                    }
+                } else {
+                    Rooter::redirectToPage("accueil");
                 }
                 break;
 
@@ -84,42 +91,50 @@ Class SuperController {
 
 
             case 'modifProduit':
-                require_once('pages/produits/ControllerProduits.php');
-                $produits = new ControllerProduits();
+                if (!empty($_SESSION['ID']) && $_SESSION['IsAdmin'] == 1) {
+                    require_once('pages/produits/ControllerProduits.php');
+                    $produits = new ControllerProduits();
 
-                if (!empty($_GET['prod'])) {
-                    $idprod = $_GET['prod'];
+                    if (!empty($_GET['prod'])) {
+                        $idprod = $_GET['prod'];
 
-                    if (!empty($_FILES['image'])) {
-                        $produits->modifImgProduit($_FILES['image'], $idprod);
+                        if (!empty($_FILES['image'])) {
+                            $produits->modifImgProduit($_FILES['image'], $idprod);
+                        }
+                        if (!empty($_POST['info']) && !empty($_POST['quoi'])) {
+                            $produits->modifProduit($_POST['info'], $_POST['quoi'], $idprod);
+                        }
+                        Rooter::redirectToPage("produits&prod=$idprod");
                     }
-                    if (!empty($_POST['info']) && !empty($_POST['quoi'])) {
-                        $produits->modifProduit($_POST['info'], $_POST['quoi'], $idprod);
-                    }
-                    Rooter::redirectToPage("produits&prod=$idprod");
                 }
                 break;
 
 
             case 'deleteProduit':
-                if (!empty($_GET['prod'])) {
+                if (!empty($_SESSION['ID']) && !empty($_GET['prod'])) {
                     require_once('pages/produits/ControllerProduits.php');
                     $produits = new ControllerProduits();
 
                     $produits->deleteProduit($_GET['prod']);
                     Rooter::redirectToPage('catalogue');
+                } else {
+                    Rooter::redirectToPage("catalogue");
                 }
                 break;
 
 
             case 'créationProduit':
-                require_once('pages/produits/ControllerProduits.php');
-                $produits = new ControllerProduits();
+                if (!empty($_SESSION['ID']) && $_SESSION['IsAdmin'] == 1) {
+                    require_once('pages/produits/ControllerProduits.php');
+                    $produits = new ControllerProduits();
 
-                $produits->includeViewCréaProduit();
-                if (!empty($_FILES['image']) && !empty($_POST['nom']) && !empty($_POST['description']) && !empty($_POST['prix']) && !empty($_POST['stock'])) {
-                    $produits->addProduit($_FILES['image'], $_POST['nom'], $_POST['description'], $_POST['prix'], $_POST['stock']);
-                    Rooter::redirectToPage('catalogue');
+                    $produits->includeViewCréaProduit();
+                    if (!empty($_FILES['image']) && !empty($_POST['nom']) && !empty($_POST['description']) && !empty($_POST['prix']) && !empty($_POST['stock'])) {
+                        $produits->addProduit($_FILES['image'], $_POST['nom'], $_POST['description'], $_POST['prix'], $_POST['stock']);
+                        Rooter::redirectToPage('catalogue');
+                    }
+                } else {
+                    Rooter::redirectToPage("catalogue");
                 }
                 break;
 
@@ -141,31 +156,37 @@ Class SuperController {
                     require_once('pages/profil/ControllerProfil.php');
                     $profil = new ControllerProfil();
 
-                    $profil->modifProfil($_SESSION['ID'], $_POST['info'], $_POST['type']);
-                    Rooter::redirectToPage('profil');
+                    $modif = $profil->modifProfil($_SESSION['ID'], $_POST['info'], $_POST['type']);
+                    if ($modif) {
+                        Rooter::redirectToPage('profil');
+                    } else {
+                        Rooter::redirectToPage('profil&err=1');
+                    }
                 }
                 break;
 
 
             case 'deleteProfil':
-                if (!empty($_GET['delete'])) {
+                if (!empty($_SESSION['ID']) && !empty($_GET['delete'])) {
                     require_once('pages/profil/ControllerProfil.php');
                     $profil = new ControllerProfil();
 
                     $profil->deleteUser($_GET['delete']);
                     Rooter::redirectToPage('profil');
+                } else {
+                    Rooter::redirectToPage('accueil');
                 }
                 break;
 
 
             case 'panier':
-                if ($_SESSION['IsAdmin'] == 0) {
+                if (!empty($_SESSION['ID']) && $_SESSION['IsAdmin'] == 0) {
                     require_once ('pages/panier/ControllerPanier.php');
                     $panier = new ControllerPanier();
 
                     $panier->includePanier();
                 } else {
-                    Rooter::redirectToPage('catalogue');
+                    Rooter::redirectToPage('accueil');
                 }
                 break;
 
@@ -177,12 +198,14 @@ Class SuperController {
 
                     $panier->addPanier($_POST['quantité'], $_POST['produit'], $_SESSION['ID']);
                     Rooter::redirectToPage('catalogue');
+                } else {
+                    Rooter::redirectToPage('accueil');
                 }
                 break;
 
 
             case 'deletePanier':
-                if (!empty($_GET['del'])) {
+                if (!empty($_SESSION['ID']) && !empty($_GET['del'])) {
                     require_once ('pages/panier/ControllerPanier.php');
                     $panier = new ControllerPanier();
 
@@ -193,7 +216,7 @@ Class SuperController {
 
 
             case 'commander':
-                if ($_SESSION['IsAdmin'] == 0) {
+                if (!empty($_SESSION['ID']) && $_SESSION['IsAdmin'] == 0) {
                     require_once ('pages/commandes/ControllerCommandes.php');
                     $commande = new CommandesDAO();
 
@@ -203,72 +226,114 @@ Class SuperController {
                     } else {
                         Rooter::redirectToPage('panier&err=1');
                     }
+                } else {
+                    Rooter::redirectToPage('catalogue');
                 }
                 break;
 
 
             case 'listeUtilisateurs':
-                require_once('pages/utilisateurs/ControllerUtilisateurs.php');
-                $listUsers = new ControllerUtilisateurs();
+                if (!empty($_SESSION['ID']) && $_SESSION['IsAdmin'] == 1) {
+                    require_once('pages/utilisateurs/ControllerUtilisateurs.php');
+                    $listUsers = new ControllerUtilisateurs();
 
-                $listUsers->includeViewUtilisateur();
+                    $listUsers->includeViewUtilisateur();
+                }
                 break;
 
 
             case 'deleteUser':
-                if (!empty($_GET['id'])) {
+                if (!empty($_SESSION['ID']) && !empty($_GET['id'])) {
                     require_once('pages/utilisateurs/ControllerUtilisateurs.php');
                     $listUsers = new ControllerUtilisateurs();
 
                     $listUsers->deleteUser($_GET['id']);
                     Rooter::redirectToPage('listeUtilisateurs');
+                } else {
+                    Rooter::redirectToPage('accueil');
                 }
                 break;
 
             case 'catégories':
-                require_once('pages/catégories/ControllerCatégories.php');
-                $catégories = new ControllerCatégories();
+                if (!empty($_SESSION['ID']) && $_SESSION['IsAdmin'] == 1) {
+                    require_once('pages/catégories/ControllerCatégories.php');
+                    $catégories = new ControllerCatégories();
 
-                $catégories->includeViewCatégories();
+                    $catégories->includeViewCatégories();
+                } else {
+                    Rooter::redirectToPage('accueil');
+                }
                 break;
 
 
             case 'modifCatégorie':
+                if (!empty($_SESSION['ID']) && $_SESSION['IsAdmin'] == 1) {
+                    require_once('pages/catégories/ControllerCatégories.php');
+                    $modifCatégories = new ControllerCatégories();
 
-                require_once('pages/catégories/ControllerCatégories.php');
-                $modifCatégories = new ControllerCatégories();
-
-                if (!empty($_GET['caté'])) {
-                    $modifCatégories->includeViewModifCatégorie();
-                }
-                if (!empty($_POST['nom']) && !empty($_POST['description']) && !empty($_POST['sousCaté']) && !empty($_POST['idCaté']) && !empty($_POST['idSousCaté'])) {
-                    $modifCatégories->modifCatégorie($_POST['nom'], $_POST['description'], $_POST['sousCaté'], $_POST['idCaté'], $_POST['idSousCaté']);
-                    Rooter::redirectToPage("catégories");
+                    if (!empty($_GET['caté'])) {
+                        $modifCatégories->includeViewModifCatégorie();
+                    }
+                    if (!empty($_POST['nom']) && !empty($_POST['description']) && !empty($_POST['sousCaté']) && !empty($_POST['idCaté']) && !empty($_POST['idSousCaté'])) {
+                        $modifCatégories->modifCatégorie($_POST['nom'], $_POST['description'], $_POST['sousCaté'], $_POST['idCaté'], $_POST['idSousCaté']);
+                        Rooter::redirectToPage("catégories");
+                    }
+                } else {
+                    Rooter::redirectToPage("accueil");
                 }
                 break;
 
 
             case 'deleteCatégorie':
-                if (!empty($_GET['id'])) {
+                if (!empty($_SESSION['ID']) && !empty($_GET['id'])) {
                     require_once('pages/catégories/ControllerCatégories.php');
                     $catégories = new ControllerCatégories();
 
                     $catégories->deleteCatégorie($_GET['id']);
                     Rooter::redirectToPage("catégories");
+                } else {
+                    Rooter::redirectToPage("catalogue");
                 }
                 break;
 
 
             case 'créationCatégorie':
-                require_once('pages/catégories/ControllerCatégories.php');
-                $créaCatégories = new ControllerCatégories();
+                if (!empty($_SESSION['ID']) && $_SESSION['IsAdmin'] == 1) {
+                    require_once('pages/catégories/ControllerCatégories.php');
+                    $créaCatégories = new ControllerCatégories();
 
-                $créaCatégories->includeViewCréationCatégorie();
+                    $créaCatégories->includeViewCréationCatégorie();
 
-                if (!empty($_POST['nom']) && !empty($_POST['description']) && !empty($_POST['sousCaté'])) {
-                    $créaCatégories->créationCatégorie($_POST['nom'], $_POST['description'], $_POST['sousCaté']);
-                    Rooter::redirectToPage("catégories");
+                    if (!empty($_POST['nom']) && !empty($_POST['description']) && !empty($_POST['sousCaté'])) {
+                        $créaCatégories->créationCatégorie($_POST['nom'], $_POST['description'], $_POST['sousCaté']);
+                        Rooter::redirectToPage("catégories");
+                    }
+                } else {
+                    Rooter::redirectToPage("catalogue");
                 }
+                break;
+
+
+            case 'messages':
+                if (!empty($_SESSION['ID'])) {
+                    require_once('pages/messages/ControllerMessages.php');
+                    $messages = new ControllerMessages();
+
+                    if (!empty($_GET['viewed']) && !empty($_GET['id'])) {
+                        $messages->changeStatutMessage($_GET['id'], $_GET['viewed']);
+                        Rooter::redirectToPage('messages');
+                    } 
+                    else {
+                        $messages->includeViewMessages();
+                    }
+                } 
+                else {
+                    Rooter::redirectToPage('accueil');
+                }
+                break;
+
+            default:
+                Rooter::redirectToPage("accueil");
                 break;
         }
     }
